Fix holes not combining and add NA cell
Fixed adjacent holes not combining because chart container does not
update. For this, NA cells are added during times that no event happens.
Furthermore, add features for coalescing adjacent holes that are more
than two, track states when you simulate backwards, and move to the next
block instead of starting at the top.